requirements.txt
astroid==1.6.0
bleach==2.1.2
coverage==4.4.2
decorator==4.2.1
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11
django-autoslug==1.9.3
django-cors-headers==2.1.0
django-coverage-plugin==1.3
django-debug-toolbar==1.6
django-taggit==0.22.2
django-test-plus==1.0.12
djangorestframework==3.6.3
factory-boy==2.7.0
fake-factory==9999.9.9
flake8==2.5.4
html5lib==1.0.1
ipdb==0.10.0
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.2.15
jedi==0.11.1
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.4.0
matplotlib==2.2.2
parso==0.1.1
pep8==1.7.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
psutil==5.4.6
ptyprocess==0.5.2
py==1.5.2
pyflakes==1.0.0
Pygments==2.2.0
pylint==1.7.4
pylint-common==0.2.5
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
pytest==3.2.3
pytest-cov==2.5.1
pytest-django==2.9.1
pytest-sugar==0.7.1
python-decouple==3.1
pytz==2017.3
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
static3==0.7.0
termcolor==1.1.0
traitlets==4.3.2
uWSGI==2.0.15
Unipath==1.1
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==3.3.1
wrapt==1.10.11
markdown2